requirements.txt
altgraph==0.17.3
APScheduler==3.10.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
Deprecated==1.2.13
exceptiongroup==1.1.1
future==0.18.3
h11==0.14.0
idna==3.4
jsonpath==0.82
lxml==4.9.2
macholib==1.16.2
MarkupSafe==2.1.2
outcome==1.2.0
packaging==21.3
Pillow==9.5.0
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
pyparsing==3.0.9
PySocks==1.7.1
pytz==2023.3
rdbtools==0.1.15
redis==4.3.4
requests==2.30.0
requests-toolbelt==1.0.0
robotframework==6.0.2
robotframework-pythonlibcore==4.1.2
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==6.0.0
selenium==4.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.0
trio-websocket==0.10.2
tzlocal==5.0.1
urllib3==1.26.15
wechatsogou==4.5.4
Werkzeug==0.16.0
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1

openai==0.27.6
langchain==0.0.165
